+++ python/sciserver/tests/conftest.py
@@ -6,7 +6,7 @@
 # @Author: Brian Cherinka
 # @Date:   2017-08-04 14:24:44
 # @Last modified by:   Brian Cherinka
-# @Last Modified time: 2017-08-30 11:59:12
+# @Last Modified time: 2017-09-10 23:17:25
 
 from __future__ import print_function, division, absolute_import
 import pytest
@@ -18,6 +18,7 @@ from sciserver.casjobs import CasJobs
 from sciserver.scidrive import SciDrive
 from sciserver.skyserver import SkyServer
 from sciserver.skyquery import SkyQuery
+from sciserver.compute import Compute
 
 
 userinfo = [('testuser', 'testpass')]
@@ -71,6 +72,13 @@ def skquery():
     skquery = SkyQuery()
     yield skquery
     skquery = None
+
+
+@pytest.fixture(scope='session')
+def comp():
+    comp = Compute()
+    yield comp
+    comp = None
 
 
 @pytest.fixture(scope='session')